# packages/backend-core/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/webhooks": {
      "get": {
        "operationId": "WebhooksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "post": {
        "operationId": "WebhooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/webhooks/{id}": {
      "patch": {
        "operationId": "WebhooksController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchWebhookBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Munin",
    "description": "HTTP surface area for the Munin platform. Authenticate with a bearer token (admin API key or delegated end-user token) or a session cookie.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "munin_session"
      }
    },
    "schemas": {
...
      "CreateWebhookBody": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "url"
        ]
      },
      "PatchWebhookBody": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}